rename the dir to background
