+++ pkg/background/common/util.go
