+++ pkg/background/generate/cleanup/resource.go
