+++ pkg/background/generate/report.go
