+++ pkg/background/generate/resource.go
