+++ pkg/background/generate/validate.go
